pyproject.toml
[tool.poetry]
name = "earnings-call-transcriber"
version = "0.1.0"
description = "End-to-end pipeline for discovering, transcribing, and analyzing earnings call audio across Asian markets"
authors = []
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# Web framework
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}

# Database
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
pgvector = "^0.3.0"

# Task queue
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.2.0"

# Configuration
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"

# HTTP & scraping
httpx = "^0.28.0"
beautifulsoup4 = "^4.12.0"
pyyaml = "^6.0"
selenium = "^4.27.0"
yt-dlp = "^2024.12.0"

# Transcription
faster-whisper = "^1.0.0"
openai = "^1.58.0"

# Speaker diarization
pyannote-audio = "^3.3.0"
# Pin torchaudio 2.6.0 — pyannote 3.x needs AudioMetaData API removed in 2.7+
torchaudio = "2.6.0"
torch = "2.6.0"
# Pin huggingface_hub <1.0 — pyannote 3.x passes use_auth_token kwarg dropped in 1.0
huggingface-hub = ">=0.20.0,<1.0.0"

# LLM analysis
anthropic = "^0.42.0"
jinja2 = "^3.1.0"

# Logging
structlog = "^24.4.0"

# Audio processing
pydub = "^0.25.0"
ffmpeg-python = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
ruff = "^0.8.0"
mypy = "^1.13.0"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
